+++ conf/script/src/build_system/cmd/compiler/host/get_info/version/gcc/__init__.py
@@ -1,1 +1,1 @@
-from build_system.cmd.compiler.host.get_info.version.gcc.fetch_gcc_version import fetch
+from build_system.cmd.compiler.host.get_info.version.gcc.impl import fetch
